- ping time out should use %d not %lu, cast it to int
git-svn-id: http://svn.ratbox.org/svnroot/trunk@4639 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/client.c
@@ -365,8 +365,8 @@ check_pings_list(dlink_list *list)
               cptr->flags2 |= FLAGS2_PING_TIMEOUT;
 
 	      (void)ircsprintf(scratch,
-			       "Ping timeout: %lu seconds",
-			       CurrentTime - cptr->lasttime);
+			       "Ping timeout: %d seconds",
+			       (int)(CurrentTime - cptr->lasttime));
 	      
 	      (void)exit_client(cptr, cptr, &me, scratch);
               continue;
